requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.4.1
    # via uvicorn
bcrypt==3.2.0
    # via -r requirements.in
cffi==1.14.6
    # via bcrypt
click==8.0.1
    # via uvicorn
fastapi==0.68.1
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
passlib==1.7.4
    # via -r requirements.in
pycparser==2.20
    # via cffi
pydantic==1.8.2
    # via fastapi
pyjwt==2.1.0
    # via -r requirements.in
python-dotenv==0.19.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
six==1.16.0
    # via
    #   bcrypt
    #   python-multipart
starlette==0.14.2
    # via fastapi
typing-extensions==3.10.0.2
    # via pydantic
uvicorn==0.15.0
    # via -r requirements.in
